Attempt WebChannel symbolication for from-file.
Now that we have a persistent WebChannel that can be used
for symbolication, we can use it for more than just /from-browser
profile sources.
This patch is the first step: When loading an individual profile
(not a zip) from a file, we try to connect to the WebChannel and
obtain symbols.

This also requires work on the Firefox side: https://bugzilla.mozilla.org/show_bug.cgi?id=1726281