No issue: move 'Which mobile browser?' section to the bottom.
Setting up remote profiling is really easy. However, I always found
myself hesitant to share this page because I was concerned it was
intimidating: I think it's because the "Which mobile browser?" section
is so long. As such, I moved it to the bottom and provided some default
recommendations in its place.