Adjust return addresses during profile processing.
This is a crucial piece of the puzzle for richer symbolication information.
It's also the messiest piece.

I've wanted to submit this patch for a long time, but its effects can
be hard to understand as long as the other pieces that depend on it haven't
landed. With the source view deploy preview, I can now show the effects
of this patch more clearly.

Basically, without this patch, the line numbers and inline frames of the
vast majority of frames would be wrong: Time would be spent in a function
call, but the source view would blame some other code *after* the expensive
function call. Or even code somewhere else in the function entirely.
And with inline callstacks, the extra frames around function calls would
be completely wrong and confusing.

Here's an example:

https://deploy-preview-3530--perf-html.netlify.app/public/f7rxmryzgcc00qa1r7jzxzqbcgr8szzht6nr4f8/calltree/?globalTrackOrder=0wced&hiddenGlobalTracks=0wce&hiddenLocalTracksByPid=6874-024wa~6909-0w2&localTrackOrderByPid=6874-9a01324w8~6906-10~6894-10~6901-0~6914-10~6907-10~6912-10~6876-10~6900-120~6897-10~6895-120~6913-10~6893-10~6909-120~6896-120&sourceTab=hg%3Ahg.mozilla.org%2Fmozilla-central%3Atoolkit%2Fxre%2FnsEmbedFunctions.cpp%3A9cbf4fe3f852cede86354eb884cd33d305026b17&symbolServer=https%3A%2F%2Fsymbolication.stage.mozaws.net%2F&thread=x1&timelineType=cpu-category&v=6

Here, the line `process->CleanUp();` is blamed, when in fact the time is spent
a few lines further up, in `uiMessageLoop.MessageLoop::Run();`.
This profile was symbolicated when it still had the "wrong" addresses.
After the profile upgrading, the addresses are now fixed but the line
numbers from symbolication are still wrong. By triggering re-symbolication,
the line numbers can be corrected.
Here's the same profile but after re-symbolication:

https://deploy-preview-3530--perf-html.netlify.app/public/mhf089tme967h6m4c874z7nt89h7g70y32zh5dr/calltree/?globalTrackOrder=0wced&hiddenGlobalTracks=0wce&hiddenLocalTracksByPid=6874-024wa~6909-0w2&localTrackOrderByPid=6874-9a01324w8~6906-10~6894-10~6901-0~6914-10~6907-10~6912-10~6876-10~6900-120~6897-10~6895-120~6913-10~6893-10~6909-120~6896-120&sourceTab=hg%3Ahg.mozilla.org%2Fmozilla-central%3Atoolkit%2Fxre%2FnsEmbedFunctions.cpp%3A9cbf4fe3f852cede86354eb884cd33d305026b17&symbolServer=https%3A%2F%2Fsymbolication.stage.mozaws.net%2F&thread=x1&timelineType=cpu-category&v=6

The time is now correctly attributed to `uiMessageLoop.MessageLoop::Run();`.
You can see the same effect in lots of other places.