Compute correct relative addresses for kernel addresses.
When profiling kernel code, we can end up with address values which
are larger than 2^53. Calling parseInt on a hex string with such a
large address loses precision and we end up with an imprecise address.
Then we convert that imprecise address to an incorrect relative address,
look up that relative address during symbolication, and get the wrong symbol.

This commit leaves the address in hex string form for longer,
and implements subtraction using a custom U64 class.

This might be overkill but I was having trouble integrating BigInt.