Rename getProfilesFromRawUrl to retrieveProfileForRawUrl.
"retrieve" instead of "get" so that it better reflects the fact that
it doesn't only return the profile; it kicks off the entire profile
loading process, dispatches various actions, and calls loadProfile once
the profile is available. This also is more similar to the other
"retrieveProfile*" functions it calls.

"for" instead of "from" because the URL that is passed to this function
is something like /public/HASH/?... - it's the URL of this web app, and
not the URL that the profile file needs to be fetched from.