Add activeTab information to createGeckoProfile.
createGeckoProfile is used in UrlManager.test.js which has a test
for /from-browser/?view=active-tab . At the moment, this test only
passes because it doesn't wait for the profile to be fully loaded.
Once the profile is fully loaded, we would find out that it doesn't
contain any activeTab information, and would switch to the full
profile view.
Including activeTab information in this profile means that this test
will keep passing once the URL setup flow is changed to wait for
from-browser profiles correctly.