Move the call to createBrowserConnection into UrlManager.
This means we don't need to awkwardly propagate the browserConnection
back out of retrieveProfileFromBrowser and retrieveProfileForRawUrl.

It does mean that we will call createBrowserConnection from more places.
I'm aiming to reduce the number of callsites of createBrowserConnection
in the future, but the path to get there is somewhat unclear at the moment.