Make <Root> establish the browser connection.
This moves the call to createBrowserConnection all the way to the outside.
It also removes the user agent override: We will now only wait for the
WebChannel if we're running in a browser with a Firefox userAgent.
Fixes #3638.

I've kept the tests mostly as-is. I'm planning to simplify them in
future commits.