Change requestSymbolsFromServer return value.
Rather than returning an array of promises, where each promise indicates
whether a lib was found or not, return just one promise and an array
of objects which either indicate success or failure.

This is a pure refactoring and does not change behavior.

I'm doing this because I want to add another step in the symbolication
sequence: After the Mozilla symbolication API call fails, I want to
fall back to calling the browser's querySymbolicationApi. I'd like this
fallback to make one querySymbolicationApi call per chunk. So I need a
way to await the entire chunk. That's hard to do if there are individual
promises for each library in a chunk. After this refactoring it will be
straightforward.