Add utility functions to compute line timings.
These functions are used to compute the timings that will be shown in
the source view's gutter. Each line of source code has a total time and
a self time, similar to the call tree.

These functions compute all hits for a given file, across the entire
thread. So what's displayed in the gutter is independent of the select
call node.