Add a getStackLineInfoForCallNode function.
This is similar to the other function, but it computes timings for a
specific call node only.
I'm not sure if we want to display these timings in the source view
gutter. Maybe we can have a "Follow selection" checkbox later, to toggle
between the two modes.

You could even imagine three modes:
 - Show global timings for this file
 - Show timings for time spent in the selected function, anywhere in the
   call tree
 - Show timings for time spent in this call node only

I'm planning to use this function for one specific purpose:
When a call node is double clicked, I want to open the source view for
the clicked function's source file, and scroll to the spot which has the
most line hits *for the clicked call node*.
I first tried using only the "global" timings for this smart scrolling,
but it worked pretty badly: The file would open but it would be in the
completely wrong place, because some other part of the profile spent time
in elsewhere in that file. You really want to see the part of the file
relevant to the clicked call node.