Improve preview selection performance, part 2.
Use memoize-one instead of memoize-immutable in SourceView.js.
This has much better performance than a { limit: 1 } TupleMap when the key is
a long string, because it doesn't use a map. It only does an object identity
comparison on the arguments. For long string arguments, this skips the
"rope flattening" and the hashing.

Before: https://share.firefox.dev/3pzwNOe
After: https://share.firefox.dev/31Jorvr