Use CodeMirror for syntax highlighting.
This converts the source view to use CodeMirror 6, rather than our own
virtual list implementation.
CodeMirror has the advantage that it can do syntax highlighting. Most
importantly, it does syntax highlighting in a non-blocking way, so that
the profiler stays responsive even if large files are loaded.

I tried quite hard to find other syntax highlighting solutions with
similar performance which would not require embedding an entire editor,
but was not successful. For example:

 - refractor can only highlight an entire file (blocking the UI thread),
   or you can use it to highlight one line at a time, but without taking
   into account the parsing state from the previous lines, leading to
   completely wrong syntax highlighting.
 - syntect compiled to wasm was too slow, and using it in an asynchronous
   fashion would have been very complicated.

Ultimately, for correct and non-blocking syntax highlighting, you need
to be prepared to show non-highlighted code for a short time until syntax
highlighting catches up, especially if you're scrolling to a place far
down in the file (which means you have to parse everything up to that
line). Otherwise you freeze the UI.