Remove the special handling for Renderer thread during idle thread detection (PR #3921)
We had a special handling for Renderer thread because previously we had
screenshot markers in that thread. If people choose to share the profile
with hidden threads removed, then they could also lose the screenshots
coincidentally. We moved the screnshot markers to the main thread in
https://bugzilla.mozilla.org/show_bug.cgi?id=1742722 and it is safe to
remove this handling now.