Do not show IPC markers twice for a single IPC message
Previously we were seeing duplicate messages for a single IPC message in
a thread. This was a pre-existing issue but wasn't really visible until
we started to capture IPC markers from unprofiled threads. Because now
we can actually have multiple threads in a single thread, previously we
were just failing to capture the I/O related markers. With this, change:
 - If both IPC endpoint markers (start and end) are present, we will
   only see these IPC messages in start and end markers' threads.
- If one of these are not present, we will see the intermediate markers
  in the either IPC I/O thread (if profiled) or we will see the
  intermediate markers in the main threads.