Better perf script format detection
The `perf script` format is a text format that doesn't contain a
distinctive header. Our detection was too general and was incorrectly
detecting JSON profiles as perf script profiles.

Now we bail out if the file starts with an opening brace.

I've also simplified how we treat perf script profiles with header.
Rather than skipping the header and detecting the first sample, we just
detect the header.