Reduce activity threshold for non-idle threads.
Fixes #4190.

Now we will display more minimally-active threads by default.
For the example profile in #4190, a threshold of 8% would have been
sufficient to show the thread. I decided to go down to 5%, but not for
any particular reason, just because it was a rounder number. Maybe we'll
find out that 8% would have been better after all.

Reducing this threshold won't overload the thread list, because we still
have the limit of 15 threads.