Add address-timings.js and tests.
This is a first step on the path to an assembly view.
It adds the ability to compute, per native symbol, the instructions
addresses that we encountered during profiling, along with their total
and self times.
These sample counts will be shown in the assembly view gutter, similarly
to the source view gutter.