Add nativeSymbols functionSize column.
In order to know how much assembly code to display for a native symbol
in the assembly view, we need to know how large the function is, i.e.
how many bytes of machine code in contains.

In the assembly view we want to display the entire function, not just the
assembly code up until the highest instruction address that we've seen
for that function.

The symbolication API now gives us this information, as of
https://github.com/mstange/profiler-get-symbols/commit/e8649bf0cf571cfc4844a9ae7f29e61f79afbe65 .
So we can have this information for local builds and when using samply.
The Mozilla symbolication API does not return this information yet, so
we need to support "null" values with some kind of fallback.

This commit gets the information from the symbolication result, if
present, and forwards it into the nativeSymbols table.
We also bump the processed profile version and add an upgrader.