Add an isMainThread field. (PR #4459)
This adds a boolean `isMainThread` field  to the processed format, but
not to the Gecko format.

As a side-effect, this patch makes string values in pids usable. In the
past, string pids would always cause a thread to be treated as a main
thread.

In samply I want to have pids of the form "12345.1" for when process
12345 calls exec and turns into a different process with the same pid.

Fixes #2508.