Make the CallTree store a Thread.
There wasn't really a reason not to do this, since we pass
the entire Thread to the constructor. And this makes it easier
to use the full thread for other things.

For example, for the assembly view, I'll need the frame and func tables as well,
to compute the native symbols for a call node.