Add "address timings" selectors for assembly view.
This gets the timings for the native symbol which is currently being
shown in the assembly view.

We have both a "selected thread" and a "selected call node" version.
The "selected thread" version computes thread-global timings.

Unfortunately, both selectors are named the same, just like they are for
the source view.