Change the Thread pid field to always be a string.
This resolves an ambiguity during URL parsing. Fixes #4502.

This helps with profiles from samply, which now use string PIDs such as
"1234", "1234.1", "1234.2" in order to ensure unique PIDs when a process
with PID 1234 execs into a new process with the same PID.

This patch does not properly escape string PIDs with '-' or '~' in them
during URL serialization. This is an existing issue and was already
broken before this patch.