Rename "collapse indirect recursion".
The "collapse indirect recursion" transform collapses both direct and
indirect recursion. It is a good default recursion collapsing transform.
The "collapse direct recursion" transform is less useful and will be
demoted to a secondary choice.

This commit renames things as follows:

"Collapse indirect recursion" -> "Collapse recursion"
"Collapse direct recursion" -> "Collapse direct recursion only"

We also swap the order in the context menu, and we swap the r and R
shortcut keys.

This commit also removes the unused "implementation" field from the
transform struct, and adjusts the URL encoding, bumping the URL version.
This is all in the same commit because the new URL encoding takes both
the new name and the removal of the implementation field into account.

You will see some comments in the V3 -> V4 URL upgrader. I'm unsure what
to do about them. I think we have three options for what to do with the
V4 upgrader:

 - Accept the new breakage in a few cases, by leaving it largely as is.
 - Don't accept any new breakage by making it so that it correctly
   parses V3 URLs and stringifies V4 URLs. This requires copy-pasting
   the V3 transform parsing code.
 - Break it entirely by removing most of the V4 upgrader. I'm not sure
   how many V3 URLs with focus-subtree transforms there are in circulation.