Supply cssToDeviceScale to drawCanvas.
In the past, the components which draw in device space were using
window.devicePixelRatio and relying on the assumption that this is the
same scale that ChartCanvas was using when it set the canvas backing
size and CSS size. This assumption is correct, but it still feels better
to pass the scale along explicitly, rather than assuming something about
the global state.

I'm also adding some checks so that it's obvious to readers of the
drawCanvas methods how the canvas scale is controlled.