Split the thread  per-thread selectors to be able to get markers for transforms
Before the 'filter-samples' transform, we didn't need to get the
derived markers to be able to compute any thread related pre-thread
selectors. But with 'filter-samples' transform, we need to get the
derived markers to be able to compute the transform filtered thread
because this transform needs to find the markers that match the filter
and get the interval range out of them.

So because of this, I had to split the thread related selectors so we
can get the marker selectors in between them and have the marker
selectors in the second phase of thread selectors.

Previously the dependency graph was like this:

```
thread selectors
     |
     v
stack and sample
     +
marker selectors
     |
     v
composed selectors
```

Now it's like this:

```
basic thread selectors
       |
       v
marker selectors
       |
       v
rest of the thread selectors
       |
       v
stack and sample selectors
       |
       v
composed selectors
```

We can't put these thread selectors that require markers into the
`composed` selectors because stack and sample selectors require those
selectors earlier in the pipeline and I wanted to keep the thread
selectors still in their file to keep the pipeline in a single file.