DHAT: Optimize import time for large profiles (PR #4640)
* feat: optimize DHAT import tree building algorithm

This commit changes the way in which the stackTable tree is constructed
in DHAT's import algorithm.

Previously, the algorithm would look a bit like this:
    For each PP:
        For each frame in the PP, from top to bottom:
            Loop over all existing stackTable entries, until we find the
            one we want, or go through all of them.

            If we found an entry, we are good.
            If not, create a new entry.
        [...]
    [...]

This approach, while correct, isn't optimal, since we would potentially
loop over a wide range of stackTable entries that couldn't possibly be
the one we were looking for.

With this commit, the algorithm becomes more or less like this:
    For each PP:
        For each frame in the PP, from top to bottom:
            Loop over the stackTable entries whose prefix is the same as
            the one we're looking for.

            If we found an entry, we are good.
            If not, create a new entry, also updating the list of
            entries with our prefix.
        [...]
    [...]

With this, we avoid looking over a very big number of entries whose
prefix are different from ours. The effect is that of switching from a
linear search on a list to following pointers on a tree structure.

The list of entries with the same prefix is implemented as a JS Map of
prefixes to lists of indexes into the stackTable.

With this change, big DHAT profiles take a lot less time to load (e.g. a
20MB profile would take ~2min before, now it takes ~2.5s :D).

Signed-off-by: flow <flowlnlnln@gmail.com>