Replace arrow panel shadow filter with box-shadow.
The filter was used so that the panel background could be slightly transparent.
With a transparent panel, having a box-shadow would cause a visual issue
for the arrow triangle, because the arrow triangle would also be filled
with a slightly translucent background, so the panel shadow would be
visible underneath the arrow and cause a visual disconnect.

This patch makes the panel backgrounds fully opaque. Now the opaque arrow
can fully occlude the panel shadow edge.

Using box-shadow avoids some rendering bugs in Firefox, and is better for performance.