Make arrow panels more versatile.
This PR implements some changes to the arrow panel code which make arrow
panels usable for the dropped functions panel in the call tree settings.

With the previous implementation, I encountered these issues:
 - When the source view was open, the panel would be clipped to the
   call tree bounds and could not overlap the source view. The
   overflow:auto clip from the splitter view was applying to the panel.
 - With a narrow viewport or an unusually positioned anchor element, the
   panel would sometimes be positioned partially offscreen.
 - It was tricky to enforce a maximum height on the panel in a way that
   squished the list box inside the panel.
 - The panel would reposition itself based on the size of the anchor
   element *while the panel was open*. E.g. you would click an "x" button
   to remove a dropped function, and the anchor button's label would
   change from "10 dropped functions" to "9 dropped functions", causing
   the button to shrink. Then the panel would shift slightly to the left,
   moving the "x" button out from under the mouse cursor.

This commit fixes all of these issues.

The most important change is the switch to position:fixed: Now the panel
will no longer be clipped by overflow:hidden/auto ancestor of the button
that opens the panel.
Furthermore, the panel position is only computed during opening and then
remains unchanged while the panel is open.
And there is some logic to make the panel fit onscreen (while also making
sure that the arrow still points at the right place).