Reject out-of-bounds positions.
Fixes #4763.

Mouse positions towards the right end of the canvas were getting rounded
so that deviceX == canvasPixelWidth, and then we were getting undefined
out of the array in _getCPURatioAtX.

This commit changes the rounding to use floor, which keeps these
almost-right-edge positions in-bounds. Furthermore, the bounds check from
_categoryAtDevicePixel is moved into getSampleAndCpuRatioAtClick so that
it protects _getCPURatioAtX as well.

I tried to add a test for this in SampleTooltipContents.test.js, but I
couldn't get the tooltip to appear at all in the rightmost pixel. I think
the test currently spaces out the samples so far that, even with the
smoothing, the graph fill doesn't extend far enough to the right that we
would show the tooltip.