Don't expose the inverted stackTable.
This is the first step on the long road to a faster inverted call tree.

With this change, we still build an inverted call node table. But the
thread remains untouched.
This means that the stackTable and the callNodeTable are upside down with
respect to each other when the call tree is inverted. This can be a bit
confusing.

Some code needs to match non-inverted stacks to inverted call nodes. For
these purposes, a getMatchingAncestorStackForInvertedCallNode function
is introduced.

Some address timings / line timings code is now unnecessary, specifically
the code that was computing global information (and not information about
a single call node) with special treatment for the inverted case.
The unnecessary inverted case implementation was removed.