Improve flame graph performance by caching rows.
The old code was saving time by sorting siblings only for the nodes that
were displayed based on the current (preview) range selection. This made
it cheaper to compute the flame graph for a small range, but it meant
that it had to re-do the sort every time the selection changed.

Now we do the sorting once, based on the entire call node table. This is
expensive but it is a one-time cost. Then we cache the "ordered rows" and
don't have to sort again on each range selection change.