Make CallNodeInfo an interface.
This puts the conversion functions between call node indexes and call
node paths onto the new CallNodeInfo interface. It also provides accessor
methods for the traditional members.

This will allow us to have two implementations in the future: One for the
regular call tree, and one for the inverted call tree.