Move call node path/index conversion functions into CallNodeInfo.
This makes CallNodeInfo an interface, and puts the conversion functions
between call node indexes and call node paths onto this new interface.
It also provides accessor methods for the traditional members.

This reduces the number of direct uses of the call node table.
It will allow us to have two implementations in the future: One for the
non-inverted call tree, and one for the inverted call tree.