Compute new selection earlier when inverting.
Compute the inverted selected call node in the action creator,
not in the reducer. This means we don't have to pass the call tree as
far down.

Also move the computation into the CallTree class.

And fix it to stop at nodes with heaviest self time, even if they're not
leaf nodes.