Fix test expectations.
This test was accidentally passing inverted=true in the part of itself that was
testing the non-inverted tree.
There's another part further down where it calls computeCallTreeCountsAndSummary
again for the inverted tree, and that's the call where it should pass true (and
it does so correctly). The first call should pass false, and the self time
should be in the leaves.