Move viewport by the right amount.
Dragging the stack chart horizontally moved the chart less distance
than the mouse moved, because we weren't taking the margins into
account in the calculations.
viewportLeft and viewportRight refer to the edges of the "inner" viewport,
whereas containerWidth is the "outer" width. To get the inner width, i.e.
the width that the viewport coordinates target, we need to subtract the
margins from containerWidth.