Compute timeAtStart/timeAtEnd correctly.
There was some confusion with the units and margins.
This commit also moves the computation outside the row loop, because
it only needs to be done once.

This is what was wrong in the old code:

timeAtEnd wasn't taking the right margin into account, and
the computation of timePerPixel was nonsensical.

```js
const pixelsInViewport = viewportLength * innerDevicePixelsWidth;
const timePerPixel = rangeLength / pixelsInViewport;
const timeAtStart: Milliseconds =
  rangeStart + rangeLength * viewportLeft - timePerPixel * marginLeft;
const timeAtEnd: Milliseconds = rangeStart + rangeLength * viewportRight;
```

 1. innerDevicePixelsWidth is already the pixels in the viewport.
    Multiplying it with viewportLength again computes some smaller
    meaningless value.
 2. rangeLength is the length (in milliseconds) of the full range, not of
    the range in the viewport. Dividing the full range by the pixels in
    the viewport is not meaningful. It would maybe have made some sense to
    divide the full range by "the width of the entire profile at the current
    zoom level", i.e. `innerContainerWidth / viewportLength` , but that's
    not what's happening here.
 3. timePerPixel was probably intended to be timePerDevicePixel, but now
    it's getting multiplied by marginLeft, which is in CSS pixels.

All in all, `timePerPixel` is off by `viewportLength^2 * devicePixelRatio`,
i.e. it's too big and the more you zoom in the further off it gets.
Luckily, `timePerPixel` was only used in the `timePerPixel * marginLeft`
computation. So we were extending the start of the profile unnecessarily
far to the left, i.e. we were drawing more boxes than necessary beyond
the left edge of the canvas when zoomed in.