Show the current performance metric markers in the network track
It looks like the FirstContentfulPaint and FirstContentfulComposite
markers were always broken because it was actually looking at the wrong
field. It was assuming that these markers don't have any payload, but
in reality they are text markers, so they do have payloads.

With this commit, these two markers will be available in the network
track.