Rename maxDepth to maxDepthPlusOne in many places.
Rather than talking about "0-based" and "1-based" depths, I think it would
be less confusing to always treat depths as zero-based and rename these variables.
When I was reading the stack chart code, I was initially quite confused because we
were creating an array with length maxDepth, and that seemed too short.