Add a maxLength field to the call node table, and pre-allocate the per-row arrays in computeFlameGraphRows.
We don't use computeCallNodeMaxDepthPlusOne here because that one would only
include the depth of call nodes that are used by any of the filtered samples.
But computeFlameGraphRows looks at the entire call node table, and needs to
reserve enough rows to have space even for call nodes which aren't used by
any samples.