Use _callNodeHasChildren in hasChildren().
This avoids calls to getChildren for nodes which are visible but not expanded.
The tree view calls hasChildren in order to determine whether to show the
disclosure arrow.
Avoiding the call to getChildren means that we don't have to allocate the
array and sort it. This is especially useful in the inverted tree, which
has lots of non-expanded nodes at the root level.