Use updateThreadStacks instead of replaceStackReferences.
At the time I originally wrote this code, I forgot that we had
updateThreadStacks, which is what we use everywhere else where we create
a new stack table. I only remembered replaceStackReferences which takes
two different maps for "self" frames and "prefix" frames; that distinction
is useful for return address adjusting but not for symbolication, where
we pass the same map to both arguments.
Just use updateThreadStacks instead.