Speed up collapseFunctionSubtree.
The previous implementation was creating a new stackTable, which means
that we were invalidating all the derived state that was based on the
stackTable, most importantly the call node table.
There were two benefits from creating a new stackTable: We remove
some unused stack notes for the collapsed-away descendants, and we
"resolve category conflicts", i.e. we set the category of the merged
node to the "Other" category if different categories are collapsed into
it.

The former is not all that useful - the new stack table is usually not
that much smaller.
And the latter is actually not what I want - it's better to preserve the
category of the node whose subtree gets collapsed.
Also, the old implementation was discarding the line and address
information of the collapsed node. We don't want that either.

The new implementation is much simpler: We keep the stackTable
completely unchanged, and just update any stack references (e.g. sample
stacks) to point to a different node - if they were pointing inside a
collapsed subtree, they will be changed to point to the subtree root.
This preserves line + address timings and doesn't require generating a
new call node table.

Before: https://share.firefox.dev/3Td0MvP (1.9 seconds)
After: https://share.firefox.dev/3Tf81Ds (160ms, 12x faster)