Speed up _filterThreadByFunc.
This function is used for filter-by-implementation, for example if you
switch to the JavaScript-only view.

This function was doing two things that are not necessary:

 - It was ensuring that we didn't create two sibling stack nodes with
   the same frame.
 - It was resolving category conflicts when collapsing those sibling
   nodes.

We don't need to do this - it's fine to have multiple sibling stacks
with the same frame. Only in the call node table we have to worry about
having just one sibling with the same func, but the creation of the call
node table takes care of that.

So this new implementation just maps existing stacks one-to-one to new
stacks if their function remains in the thread, and skips them otherwise.

Before: https://share.firefox.dev/471ttzj (1.1 seconds)
After: https://share.firefox.dev/3uEVnmY (300ms, 3.6x faster)