Simplify detecting whether a function recurses.
On https://share.firefox.dev/3t1PUq3 , this speeds up opening the
context menu on non-recursive functions from 420ms to 30ms (14x faster).

We now use the call node table instead of the unfiltered thread's stackTable.
The call node table is much smaller than the stack table, so it's faster
to search through.
We also don't need to look at the implementation filter when detecting
direct recursion, because the call node table is already filtered by
implementation.
And we use a Uint8Array instead of a set.

Using the raw thread's stackTable is the right thing to do in the code
where we apply the transforms as part of the derived thread pipeline.
But if we're just detecting which context menu items to show, we can use
derived state as much as we want.

Before: https://share.firefox.dev/46GjMWI 406 samples in renderContextMenuContents
After: https://share.firefox.dev/418a5iB 8 samples in renderContextMenuContents (50x faster)