Rename leaf to self in many places.
All these places now deal with non-inverted call nodes, and for those,
what we meant by "leaf" and by "self" was always the same thing.
And I prefer the word "self" because "leaf" usually means "has no children"
and that's not the case here.
We still use the word "leaf" in many parts of the documentation.