Use the suffix order in getMatchingAncestorStackForInvertedCallNode.
This function is used by getNativeSymbolsForCallNodeInverted,
getStackAddressInfoForCallNodeInverted, and
getStackLineInfoForCallNodeInverted.

This replaces a call to getStackIndexToCallNodeIndex() with a call to
getStackIndexToNonInvertedCallNodeIndex(). It also mostly removes the
use of the inverted call node table for this code. (There's still a
place that accesses callNodeInfo.getCallNodeTable().depth, but this will
be fixed in a later commit.)

We want to eliminate all callers to getStackIndexToCallNodeIndex() because
we don't want to compute a mapping from non-inverted stack index to
inverted call node index upfront.