Create inverted call nodes lazily.
This is the main commit of this PR. Now that nothing is relying on having
an inverted call node for each sample, or on having a fully-computed
inverted call node table, we can make it so that we only add entries to
the inverted call node table when we actually need a node, for example
because it was revealed in the call tree. This makes it a lot faster
to click the "Invert call stack" checkbox - before this commit, we were
computing a lot of inverted call nodes that were never shown to the user.

After this commit, CallNodeInfoInvertedImpl no longer inherits from
CallNodeInfoImpl - it is now a fully separate implementation.

This commit reduces the time spent in `getInvertedCallNodeInfo` on an
example profile (https://share.firefox.dev/411Vg2T) from 11 seconds to 718 ms.
Before: https://share.firefox.dev/3CTNApp
After: https://share.firefox.dev/492F7wl (15x faster)