Update the contribution doc to not mention about adding "Merge PR" or "PR" in front of the PR number (#4932)
Previously we were adding either "Merge PR" or "PR" in front of the PR
number but we have decided not to do that anymore in one of our project
meetigs. So now we can just leave the commit message as is, if we are
happy with using the PR title as the commit message.