Make sure that we properly add a pageData even if the favicon URL construction fails
I found out about this bug while I was writing a test, which is in the
following commit. The URL construction of `moz-extension:` protocol only
works on Firefox but it behaves differently on Chrome and node. It
doesn't fail, but it can't extract the `origin` and `hostname` properly.

With this change, we make sure that even if it fails, we properly add
the page data to the map.