Add a "Merge unaccounted native frames" transform.
This makes it easier to analyze profiles with JIT frames.

This transform gets rid of 0x12345 frames in the call tree - but only
those that haven't been accounted to a shared library, i.e. only frames
which we consider to be "probably JIT" frames. This transform does not
merge unsymbolicated frames for which we know what library they belong to.

Ideally, we wouldn't have those stray hex addresses in the profiles;
either Firefox would filter them out during stack merging, or it should
give us the correct memory ranges for JIT functions so that we can account
these addresses to the right JS function / JIT trampoline etc.
But it's not doing that correctly today.

https://bugzilla.mozilla.org/show_bug.cgi?id=1463559 is one of the bugs
about this.

For now it seems easier to just offer a transform which gets rid of these
stray JIT address frames. This new transform doesn't really clutter
anything up; users will only know about it if they right-click such a
frame / call node.

Even if Firefox starts giving us better information for JIT frames, this
transform can still be useful. For example, you could see this type of
stack frame in profiles that are gathered with samply, on applications
that use other JITs or applications which don't have Jitdump
instrumentation enabled.
Another use case would be if Firefox uses a system library that makes
use of jitting, for example a software OpenGL implementation. Those
libraries can't be expected to expose their JIT details to our profiler.