Remove the console warnings that are happening while computing the tab to thread indexes map
I added these console.warns while working on the tab selector, thinking
that it shouldn't happen frequently. But it looks like we always see
some console warnings now when we capture a new profile. This is either
happening when we reopen an old Firefox session, or there is some other
bug where we don't properly register the pages.

I filed a bug to investigate this issue: Bug 1924681

Let's remove these currently so we don't pollute the console.